--- tools/tools/config/common.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const path = require('path');
-
-const projectPath = path.resolve(__dirname, '..', '..'); // 插件目录
-
-exports.projectPath = projectPath;
